TestPicker

Documentation for TestPicker.

TestPicker.eval_in_module

TestPicker.find_and_run_test_file

TestPicker.find_related_testfile

TestPicker.get_preamble_testsets

TestPicker.get_test_files

TestPicker.is_testset

TestPicker.select_and_run_testset

TestPicker.eval_in_module — Method

Evaluate ex scoped in a Module, while activating the test environment of pkg.

TestPicker.find_and_run_test_file — Method

Run fzf with the given input and if the file is a valid one run the test with the Test environment.

TestPicker.find_related_testfile — Method

Find all test files that are close to query.

TestPicker.get_preamble_testsets — Method

Fetch all the top nodes from file and split them between a preamble and testsets.

TestPicker.get_test_files — Method

get_test_files() -> String, Vector{String}

Get full collection of test files for the given package. Return the absolute path of the test directory and the collection of test files as paths relative to it.

TestPicker.is_testset — Method

Check whether the given SyntaxNode is a @testset macro block

TestPicker.select_and_run_testset — Method

Given a fuzzy_file query and a testset query return all possible testset that match both the file and the testset names, provide a choice and execute it.