TestPicker

Documentation for TestPicker.

TestPicker.TESTENV_CACHE

TestPicker.TestsetInfo

TestPicker.build_file_testset_map

TestPicker.build_preview_arg

TestPicker.current_pkg

TestPicker.eval_in_module

TestPicker.find_and_run_test_file

TestPicker.find_related_testfile

TestPicker.get_matching_files

TestPicker.get_test_files

TestPicker.get_testsets_with_preambles

TestPicker.is_testset

TestPicker.last_leaf

TestPicker.pick_testset

TestPicker.select_and_run_testset

TestPicker.TESTENV_CACHE — Constant

We reuse the temp environment made by TestEnv to avoid trigger recompilation every time.

This can be cleared with clear_testenv_cache().

TestPicker.TestsetInfo — Type

Struct representing metadata about a testset.

TestPicker.build_file_testset_map — Method

Given a list of matched files, extract all contained testsets, and build a map that maps some displayable name for fzf to the relevant data structure.

TestPicker.build_preview_arg — Method

Build the command line function to be run by fzf to preview the relevant code lines.

TestPicker.current_pkg — Method

More flexible version of current_pkg_name from TestEnv.

TestPicker.eval_in_module — Method

Evaluate ex scoped in a Module, while activating the test environment of pkg.

TestPicker.find_and_run_test_file — Method

Run fzf with the given input and if the file is a valid one run the test with the Test environment.

TestPicker.find_related_testfile — Function

Find all test files that are close to query.

TestPicker.get_matching_files — Method

Run a non-interactive command that return all the files getting the match on the given query.

TestPicker.get_test_files — Function

get_test_files(pkg::PackageSpec = current_pkg()) -> String, Vector{String}

Get full collection of test files for the given package. Return the absolute path of the test directory and the collection of test files as paths relative to it.

TestPicker.get_testsets_with_preambles — Method

Fetch all the top nodes from file and split them between a preamble and testsets.

TestPicker.is_testset — Method

Check whether the given SyntaxNode is a @testset macro block.

TestPicker.last_leaf — Method

Fetch the last leaf from the given node to try to get the ending line of the the testset block.

TestPicker.pick_testset — Method

Call fzf again to chose which testset to evaluate. The preview is done using rg and bat.

TestPicker.select_and_run_testset — Method

Given a fuzzy_file query and a testset query return all possible testset that match both the file and the testset names, provide a choice and execute it.