TestPicker

Documentation for TestPicker.

TestPicker.TESTENV_CACHE

TestPicker.TESTMODE_TRIGGER

TestPicker.EvalTest

TestPicker.TestsetInfo

TestPicker.build_file_testset_map

TestPicker.clean_results_file

TestPicker.clean_source

TestPicker.create_repl_test_mode

TestPicker.current_pkg

TestPicker.eval_in_module

TestPicker.fzf_testfile

TestPicker.fzf_testset

TestPicker.get_matching_files

TestPicker.get_preview_dimension

TestPicker.get_test_files

TestPicker.get_testsets_with_preambles

TestPicker.identify_query

TestPicker.is_preamble

TestPicker.is_testnode

TestPicker.pick_testset

TestPicker.prepend_ex

TestPicker.preview_content

TestPicker.remove_ansi

TestPicker.run_test_file

TestPicker.save_test_results

TestPicker.select_test_files

TestPicker.separator

TestPicker.testnode_symbols

TestPicker.visualize_test_results

TestPicker.TESTENV_CACHE — Constant

We reuse the temp environment made by TestEnv to avoid trigger recompilation every time.

This can be cleared with clear_testenv_cache().

TestPicker.TESTMODE_TRIGGER — Constant

Trigger key to get into test mode.

TestPicker.EvalTest — Type

Struct containing a ran object, either a testset or a file.

TestPicker.TestsetInfo — Type

Struct representing metadata about a testset.

TestPicker.build_file_testset_map — Method

Given a list of matched files, extract all contained testsets, and build a map that maps some displayable name for fzf to the relevant data structure.

TestPicker.clean_results_file — Method

This empty the file before appending new results.

TestPicker.clean_source — Method

Obtain the source from the LineNumberNode.

TestPicker.create_repl_test_mode — Method

Create an additional repl mode for testing to the given repl.

TestPicker.current_pkg — Method

More flexible version of current_pkg_name from TestEnv.

TestPicker.eval_in_module — Method

Evaluate ex scoped in a Module, while activating the test environment of pkg.

TestPicker.fzf_testfile — Method

Run fzf with the given input and if the file is a valid one run the test with the Test environment.

TestPicker.fzf_testset — Method

Given a fuzzy_file query and a testset query return all possible testset that match both the file and the testset names, provide a choice and execute it.

TestPicker.get_matching_files — Method

Run a non-interactive command that return all the files getting the match on the given query.

TestPicker.get_preview_dimension — Function

Utility function to adapt the size of the text width and line position.

TestPicker.get_test_files — Function

get_test_files(pkg::PackageSpec = current_pkg()) -> String, Vector{String}

Get full collection of test files for the given package. Return the absolute path of the test directory and the collection of test files as paths relative to it.

TestPicker.get_testsets_with_preambles — Method

Fetch all the nodes from the given file and for each testset (included nested ones) collect all preamble statements (see is_preamble).

TestPicker.identify_query — Method

Identify the type of query based on the input.

TestPicker.is_preamble — Method

Check if a statement qualifies as a preamble.

TestPicker.is_testnode — Function

Check whether the given SyntaxNode is a @testset macro block.

TestPicker.pick_testset — Method

Call fzf again to chose which testset to evaluate. The preview is done using bat.

TestPicker.prepend_ex — Method

Create a new block by either preprending a new_line to an existing block or creating a new block with the two expresssions.

TestPicker.preview_content — Method

We connect the error with the backtrace to be previewed.

TestPicker.remove_ansi — Method

File names come with ansi characters and break stuff...

TestPicker.run_test_file — Method

Build and evaluate the expression for the given test file.

TestPicker.save_test_results — Method

Append test results from the given testset.

TestPicker.select_test_files — Function

Find all test files that are close to query.

TestPicker.separator — Method

Separator used by fzf to distinguish the different data components.

TestPicker.testnode_symbols — Method

Fetch all potential test block that can be fetched from the testset parsing. Other nodes than @testset can be added (coma separated) via ENV["TESTPICKER_NODES"].

TestPicker.visualize_test_results — Function

This creates a loop on the latest evaluated test file. It shows a list of the tests that errored and failed with a preview of their stacktrace. You can edit the selected test with Ctrl+e or inspect the stacktrace for errored tests with Enter. It is also possible to inspect the stacktrace as a list with a preview of the source when possible and Ctrl+e edit the source of the current trace.